server/academic-service/requirements.txt
django>=5.0
djangorestframework>=3.14
django-cors-headers
psycopg2-binary
python-dotenv
PyJWT>=2.8